examples/essbase-datahub-connector/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "essbase-datahub-connector"
version = "0.1.0"
description = "Extract metadata from Oracle Essbase and emit to DataHub"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Your Name", email = "you@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["essbase", "datahub", "metadata", "lineage", "etl", "olap"]

dependencies = [
    "requests>=2.28.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "acryl-datahub>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "types-requests",
    "types-PyYAML",
]

[project.scripts]
essbase-datahub = "essbase_datahub.cli:main"

[project.urls]
Homepage = "https://github.com/yourorg/essbase-datahub-connector"
Documentation = "https://github.com/yourorg/essbase-datahub-connector#readme"
Repository = "https://github.com/yourorg/essbase-datahub-connector"
Issues = "https://github.com/yourorg/essbase-datahub-connector/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=essbase_datahub --cov-report=term-missing"
